pyproject.toml
[project]
name = "zenconnect"
version = "0.1.0"
description = "Zendesk AI Agent Service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]",
    "asyncpg",
    "arq",
    "httpx",
    "pydantic-settings",
    "structlog"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-httpx",
    "anyio",
    "ruff",
    "mypy"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true
